requirements.txt
Pillow==8.1.0
pyocr==0.8

